% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn.R
\name{knn}
\alias{knn}
\title{K nearest neighbors alogrithm}
\usage{
knn(x.mat, y.vec, testx.vec, max.neighbors)
}
\arguments{
\item{x.mat}{numeric train feature matrix [ n x p ]}

\item{y.vec}{numeric train label vector [n], either all 
0/1 for binary classification, or other real numbers for
regression (multi-class classification not supported).}

\item{testx.vec}{numeric test feature vector [p]}

\item{max.neighbors}{scalar integer, max number of neighbors}
}
\value{
numeric vector of size max.neighbors, predictions
from 1 to max.neighbors.
}
\description{
An R function that wraps the nearest neighbor C++ code
}
\examples{
data(zip.train, package = "ElemStatLearn")
i01 <- which(zip.train[,1] \%in\% c(0,1))
train.i <- i01[1:5]
test.i <- i01[6]
x <- zip.train[train.i, -1]
y <- zip.train[train.i, 1]
testx <- zip.train[test.i, -1]
knn(x, y, testx, 3)
zip.train[test.i]
}
